backend/requirements.txt
Django==4.0
djangorestframework==3.13.1
django-cors-headers==3.11.0
djoser==2.1.0
djangorestframework-simplejwt<5.0.0,>=4.3.0

